refactor preheat model
Move the notification to preheat models
