+++ src/pkg/p2p/preheat/models/notification/notification.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package models
+package notification
 
 import (
 	"time"
